NPC missions clerk
Beginning a missions clerk who hides when an enemy is around.
Added a toggle for the poison dart identifier above teammates.
When clicking a missions clerk, default action is to hit them with
hands.

Fixed issue where a door would begin to close but you could still walk
into it.